+++ hosted-webauthn-for-logged-out-users/pages/login.html
@@ -58,6 +58,9 @@
           <div class="row">
             Dont have a passkey enrolled? Enroll <a href="/pages/register.html">here</a>
           </div>
+          <div class="row">
+            Want to create and sign a transaction? Click <a href="/pages/transaction.html">here</a>
+          </div>
         </div>
       </main>
     </div>
