+++ hosted-webauthn-for-logged-out-users/pages/register-complete.html
@@ -31,7 +31,7 @@
       <main>
         <div class="content column centered gap">
           <img src="/shield.svg" alt="Logged in" />
-          <div>Returned Data</div>
+          <div>Registration response data</div>
           <textarea disabled id="tokenData" class="dataTextArea"></textarea>
         </div>
       </main>
