--- hosted-webauthn-transaction-signing/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@ciam-expressjs-vanilla-samples/hosted-webauthn-transaction-signing",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "start": "vite --debug --host"
-  },
-  "dependencies": {
-    "dotenv": "16.0.3",
-    "express": "~4.18.2",
-    "express-session": "^1.17.3",
-    "morgan": "~1.10.0",
-    "node-fetch": "^3.3.0",
-    "@ciam-expressjs-vanilla-samples/shared": "1.0.0"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.17",
-    "sass": "1.58.3",
-    "vite": "^4.2.3",
-    "vite-plugin-mix": "0.4.0"
-  }
-}
